update global state
