+++ src/App.jsx
@@ -6,9 +6,10 @@ import { useSession } from '@inrupt/solid-ui-react';
 // Custom Hook Imports
 import { useRedirectUrl } from './hooks';
 // Component Imports
-import { SignedInUserContextProvider } from './contexts';
-import LandingPage from './routes/LandingPage';
-import Home from './routes/Home';
+import GlobalDataContextProvider from './contexts/GlobalDataContext';
+import Layout from './layouts/Layout';
+import AppRoutes from './AppRoutes';
+
 
 /**
  * @typedef {import("./typedefs").userListObject} userListObject
@@ -42,13 +43,13 @@ const App = () => {
     if (session.info.isLoggedIn) localStorage.setItem('loggedIn', true);
   }, [session.info.isLoggedIn]);
 
-  return session.info.isLoggedIn ? (
-    <SignedInUserContextProvider>
-      <Home />
-    </SignedInUserContextProvider>
-  ) : (
-    <LandingPage />
-  );
+  return (
+      <Layout ariaLabel="Home Page">
+        <GlobalDataContextProvider>
+          <AppRoutes />
+        </GlobalDataContextProvider>
+      </Layout>
+  )
 };
 
 export default App;
